% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{AUC}
\alias{AUC}
\title{AUC}
\usage{
AUC(Y, Y_hat)
}
\arguments{
\item{Y}{Ground truth numeric vector.}

\item{Y_hat}{Predicted Labels numeric vector.}
}
\value{
A numeric value corresponding to the AUC of binary classification problem
}
\description{
Returns the Area Under the Curve for a binarry classification problem.
}
\examples{
Y = sample(x = c(1,2), size = 10, replace = TRUE)
Y_hat = sample(x = c(1,2), size = 10, replace = TRUE)
AUC(Y = Y, Y_hat = Y_hat)
}
