% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{BinClassMetrics}
\alias{BinClassMetrics}
\title{BinClassMetricsNames}
\usage{
BinClassMetrics(Y, Y_hat, MetricsNames = BinClassMetricsNames())
}
\arguments{
\item{Y}{Ground truth numeric vector.}

\item{Y_hat}{Predicted Labels numeric vector.}

\item{MetricsNames}{can be found at BinClassMetricsNames()}
}
\value{
A binaryResultList with results
}
\description{
Returns a binaryResultList with binary classificaiton metrics.
}
\examples{
Y = sample(x = c(1,2), size = 10, replace = TRUE)
Y_hat = sample(x = c(1,2), size = 10, replace = TRUE)
BinClassMetrics(Y = Y, Y_hat = Y_hat, MetricsNames = BinClassMetricsNames())
}
