% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{BinClassMetricsNames}
\alias{BinClassMetricsNames}
\title{Binary Classification Metrics}
\usage{
BinClassMetricsNames()
}
\value{
character string vector with all possible binary classification metrics.
}
\description{
Returns a character string vector containning all binary classification metrics.
}
\examples{
BinClassMetricsNames()
}
