% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{BinMatrix2Vector}
\alias{BinMatrix2Vector}
\title{Matrix of Dummies to Vector}
\usage{
BinMatrix2Vector(Y_matrix, levels_val = NULL)
}
\arguments{
\item{Y_matrix}{a numeric, factor or character vector.}

\item{levels_val}{levels to be considered, default is NULL for automatic levels.}
}
\value{
A numeric vector
}
\description{
Transforms a matrix of dummies into a numeric vector.
}
\examples{
Y <- sample(c(1,4,3), size = 100, replace = TRUE)
Y_matrix <- factor2BinMatrix(Y)
BinMatrix2Vector(Y_matrix = Y_matrix)
BinMatrix2Vector(Y_matrix = Y_matrix, levels_val = unique(Y))
BinMatrix2Vector(Y_matrix = Y_matrix, levels_val = c('a', 'e', 'i'))
}
