% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{CreateAlgo}
\alias{CreateAlgo}
\title{Create Algorithm}
\usage{
CreateAlgo(algoName, algoFun, task, paramList)
}
\arguments{
\item{algoName}{A character string that represents the algorithm name.}

\item{algoFun}{A function class object.}

\item{task}{A character string vector, cointaning 'MultClass' or/and 'BinClass' and/or 'Regression'.}

\item{paramList}{A list of all parameters and their values to be tested, NULL if none.}
}
\value{
An object of class algorithm.
}
\description{
It is an auxiliary function to help creating new algorithms in the package standarts.
}
\examples{
algoName <- 'myAlgo'
algoFun <- function(X_train, Y_train, X_test, param1){
 set.seed(param1)
 sample(Y_train, size = nrow(X_test), replace = TRUE)}
task <- c('MultClass', 'BinClass')
paramList = list(param1 = 1:3)
CreateAlgo(algoName = algoName, algoFun = algoFun, task = task, paramList = paramList)
}
