% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{CreateDataSet}
\alias{CreateDataSet}
\title{Create DataSet}
\usage{
CreateDataSet(X, Y, Name, type, task)
}
\arguments{
\item{X}{A Matrix.}

\item{Y}{A numeric vector of classes or values.}

\item{Name}{A character string, as dataset name.}

\item{type}{A character string, the types of values for X (numeric or integer).}

\item{task}{A character string vector of task to be performed, check GetPossibleTasks().}
}
\value{
A DataSet object type.
}
\description{
Creates a DataSet object type, to be used with the models provided with this package
}
\examples{
X <- as.matrix(cbind(runif(n = 100), runif(n = 100)))
Y <- sample(x = c(1, 2), size = 100, replace = TRUE)
Name <- 'randomData'
type <- 'numeric'
task <- 'BinClas'
newData <- CreateDataSet(X = X, Y = Y, Name = Name, type = type, task = task)
}
