% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{factor2BinMatrix}
\alias{factor2BinMatrix}
\title{Factor Vector to Binary matrix of dummies}
\usage{
factor2BinMatrix(Y, levels_val = NULL)
}
\arguments{
\item{Y}{a numeric, factor or character vector.}

\item{levels_val}{levels to be considered, default is NULL for automatic levels.}
}
\value{
matrix
}
\description{
Transforms a factor vector into a binary matrix of dummies.
}
\examples{
Y <- sample(c(1,2,3), size = 100, replace = TRUE)
factor2BinMatrix(Y)
factor2BinMatrix(as.factor(Y))
factor2BinMatrix(Y, levels_val = 1:5)
factor2BinMatrix(as.factor(Y), levels_val = 1:5)
}
