% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{GetAllBinClassAlgo}
\alias{GetAllBinClassAlgo}
\title{Generates all Binary Classification Alogrithms}
\usage{
GetAllBinClassAlgo()
}
\value{
Returns a list with all Binary Classification Alogrithms
}
\description{
It is an auxiliary that allows to load all of the packages binary classification alogrithms.
}
\examples{
GetAllBinClassAlgo()
}
