% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{GetAllMultClassAlgo}
\alias{GetAllMultClassAlgo}
\title{Generates all MultClass Classification Alogrithms}
\usage{
GetAllMultClassAlgo()
}
\value{
Returns a list with all MultClass Classification Alogrithms
}
\description{
It is an auxiliary that allows to load all of the packages multclass classification alogrithms.
}
\examples{
GetAllMultClassAlgo()
}
