% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{GetAllRegressionAlgo}
\alias{GetAllRegressionAlgo}
\title{Generates all Regression Alogrithms}
\usage{
GetAllRegressionAlgo()
}
\value{
Returns a list with all regression Alogrithms
}
\description{
It is an auxiliary that allows to load all of the packages regression alogrithms.
}
\examples{
GetAllRegressionAlgo()
}
