% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{GetData}
\alias{GetData}
\title{Get DataSet by name}
\usage{
GetData(datasetName, seed, splitPerc)
}
\arguments{
\item{datasetName}{A character string, as DataSet name.}

\item{seed}{For the traint and test split.}

\item{splitPerc}{Percentage for train of all dataSet, between 0 and 1.}
}
\value{
A trainTestDataSet object type.
}
\description{
Get a dataset by name
}
\examples{
GetData(datasetName = 'Iris', seed = 123, splitPerc = 0.7)
}
