% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{GetDataSetsNames}
\alias{GetDataSetsNames}
\title{Get Datasets Names}
\usage{
GetDataSetsNames(task)
}
\arguments{
\item{task}{A character string with desired task, for possible tasks check GetPossibleTasks().}
}
\value{
A character string vector with all possible datasets names.
}
\description{
Get all available datasets names
}
\examples{
GetDataSetsNames(task = 'MultClass')
GetDataSetsNames(task = 'BinClass')
GetDataSetsNames(task = 'Regression')
}
