% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{GetMetrics}
\alias{GetMetrics}
\title{Get Metrics}
\usage{
GetMetrics(task)
}
\arguments{
\item{task}{A character string vector, cointaning 'MultClass' or/and 'BinClass' and/or 'Regression'.}
}
\value{
All metrics names for that taks
}
\description{
Get all possible task names
}
\examples{
GetMetrics(task = 'MultClass')
GetMetrics(task = 'BinClass')
GetMetrics(task = 'Regression')
}
