% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tests.R
\name{GetPossibleTasks}
\alias{GetPossibleTasks}
\title{Get Possible Tasks}
\usage{
GetPossibleTasks()
}
\value{
A character string vector with all possible task names.
}
\description{
Get all possible task names
}
\examples{
GetPossibleTasks()
}
