% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Avaliation.R
\name{hypotessesTesting}
\alias{hypotessesTesting}
\title{Reduce Results}
\usage{
hypotessesTesting(resultTable, testingMetric, method = "friedman",
  reductionMethod = "mean", reductionOrder = "max", alpha = 0.05)
}
\arguments{
\item{resultTable}{Table resulted from a MLAT::RunTests() call}

\item{testingMetric}{A decision metric, must have been selected in MLAT::RunTests() function call.}

\item{reductionMethod}{A reduction method, can be 'mean' or 'median'.}

\item{reductionOrder}{The reduction Order of selected values, max or min.}

\item{alpha}{Test significance level.}
}
\value{
A list of type hypotesesTest.
}
\description{
Reduce Results

Performs Statistical Tests
}
\examples{
cmpTestsFuncsList <- GetAllMultClassAlgo()
task <- 'MultClass'
dataSetNames <- c('Iris', 'PimaIndiansDiabetes')
myResult <- RunTests(cmpTestsFuncsList = GetAllMultClassAlgo(),
                     task = 'MultClass',
                     dataSetNames = c('Iris', 'PimaIndiansDiabetes'))
reducedTable <- ReduceResults(resultTable = myResult,
                             reductionMetric = 'Accuracy Macro',
                             reductionMethod = 'mean',
                             reductionOrder = 'max' )
}
