% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{knn}
\alias{knn}
\title{K Nearest Neighbours}
\usage{
knn(X_train, Y_train, X_test, K)
}
\arguments{
\item{X_train}{A Matrix of trainning observations.}

\item{Y_train}{A numeric vector of classes or values of the trainning observations.}

\item{X_test}{A Matrix of testing observations.}

\item{K}{An integer as a parameter for the knn method.}
}
\value{
predicted labels
}
\description{
It is the K Nearest Neighbours method
}
\examples{
X <- as.matrix(cbind(runif(n = 100), runif(n = 100)))
pos <- sample(100, 70)
X_train <- X[pos, ]
X_test <- X[-pos, ]
Y_train <- as.numeric( X_train[, 1] ** 2 - X_train[, 2] > 0)
Y_test <- as.numeric(X_test[, 1] ** 2 - X_test[, 2] > 0)
K <- 5
Y_predicted <- knn(X_train = X_train, Y_train = Y_train, X_test = X_test, K = K)
print(table(Y_test, Y_predicted))
}
