% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{LogLoss}
\alias{LogLoss}
\title{LogLoss}
\usage{
LogLoss(Y, Y_hat)
}
\arguments{
\item{Y}{Ground truth numeric vector.}

\item{Y_hat}{Predicted Labels numeric vector.}
}
\value{
A numeric value corresponding to the LogLoss of binary classification problem
}
\description{
Returns the Logarithmic Loss for classification problem.
}
\examples{
Y = sample(x = c(1,2), size = 10, replace = TRUE)
Y_hat = sample(x = c(1,2), size = 10, replace = TRUE)
LogLoss(Y = Y, Y_hat = Y_hat)
}
