% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{MeanError}
\alias{MeanError}
\title{Mean Error}
\usage{
MeanError(Y, Y_hat)
}
\arguments{
\item{Y}{Ground truth numeric vector.}

\item{Y_hat}{Predicted values numeric vector.}
}
\value{
A numeric values.
}
\description{
Returns the mean error score for regression.
}
\examples{
Y <- runif(100)
Y_hat <- Y + 0.2*runif(100)
R2(Y = Y, Y_hat = Y_hat)
ExplainedVarianceScore(Y = Y, Y_hat = Y_hat)
MeanError(Y = Y, Y_hat = Y_hat)
MeanAbsoluteError(Y = Y, Y_hat = Y_hat)
MeanSquaredError(Y = Y, Y_hat = Y_hat)
RootMeanSquaredError(Y = Y, Y_hat = Y_hat)
NormalizedRootMeanSquaredError(Y = Y, Y_hat = Y_hat)
CoefficientVariationRootMeanSquaredError(Y = Y, Y_hat = Y_hat)
MedianAbsoluteError(Y = Y, Y_hat = Y_hat)
}
