% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{MLP_classification}
\alias{MLP_classification}
\title{Multilayer Perceptron}
\usage{
MLP_classification(X_train, Y_train, X_test, n_neurons, weight_decay = 0,
  max_iter = 1000)
}
\arguments{
\item{X_train}{A Matrix of trainning observations.}

\item{Y_train}{A numeric vector of classes or values of the trainning observations}

\item{X_test}{A Matrix of testing observations.}
}
\value{
predicted values
}
\description{
Performs a multilayer perceptron classification
}
\examples{
X_train <- X[pos, ]
X_test <- X[-pos, ]
Y_train <- as.numeric( X_train[, 1] ** 2 - X_train[, 2] > 0)
Y_test <- as.numeric(X_test[, 1] ** 2 - X_test[, 2] > 0)
n_neurons <- 50
Y_predicted <- MLP_classification(X_train = X_train, Y_train = Y_train, X_test = X_test, n_neurons = n_neurons)
table(Y_test, Y_predicted)
}
