% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{MultClassMetrics}
\alias{MultClassMetrics}
\title{MultClassMetrics}
\usage{
MultClassMetrics(Y, Y_hat, MetricsNames)
}
\arguments{
\item{Y}{Ground truth numeric vector.}

\item{Y_hat}{Predicted Labels numeric vector.}

\item{MetricsNames}{Metrics names, avilable can be found with MultiClassMetricsNames() .}
}
\value{
A multiClassResultList with results
}
\description{
Returns a multiClassResultList with multi class classificaiton metrics.
}
\examples{
Y = sample(x = c(1,2, 3), size = 20, replace = TRUE)
Y_hat = sample(x = c(1, 2, 3), size = 20, replace = TRUE)
MultClassMetrics(Y = Y, Y_hat = Y_hat, MetricsNames = MultiClassMetricsNames())
}
