% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{MultiClassMetricsNames}
\alias{MultiClassMetricsNames}
\title{Multi Class Classification Metrics}
\usage{
MultiClassMetricsNames()
}
\value{
character string vector with all possible multi class classification metrics.
}
\description{
Returns a character string vector containning all multi class classification metrics.
}
\examples{
MultiClassMetricsNames()
}
