% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Avaliation.R
\name{ReduceResults}
\alias{ReduceResults}
\title{Reduce Results}
\usage{
ReduceResults(resultTable, reductionMetric, reductionMethod,
  reductionOrder = "max")
}
\arguments{
\item{resultTable}{Table resulted from a MLAT::RunTests() call}

\item{reductionMetric}{A decision metric, must have been selected in MLAT::RunTests() function call.}

\item{reductionMethod}{A reduction method, can be 'mean' or 'median'.}

\item{reductionOrder}{The reduction Order of selected values, max or min.}
}
\value{
A reduced Table of type testResultTable .
}
\description{
Reduce Results

Reduces result table by parameter
}
\examples{
cmpTestsFuncsList <- GetAllMultClassAlgo()
task <- 'MultClass'
dataSetNames <- c('Iris', 'PimaIndiansDiabetes')
myResult <- RunTests(cmpTestsFuncsList = GetAllMultClassAlgo(),
                     task = 'MultClass',
                     dataSetNames = c('Iris', 'PimaIndiansDiabetes'))
reducedTable <- ReduceResults(resultTable = myResult,
                             reductionMetric = 'Accuracy Macro',
                             reductionMethod = 'mean',
                             reductionOrder = 'max' )



}
