% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{RegMetrics}
\alias{RegMetrics}
\title{Regression Metrics}
\usage{
RegMetrics(Y, Y_hat, MetricsNames)
}
\arguments{
\item{Y}{Ground truth numeric vector.}

\item{Y_hat}{Predicted values numeric vector.}

\item{MetricsNames}{Metrics names, avilable can be found with RegressionMetricsNames() .}
}
\value{
A regResultList with results
}
\description{
Returns a regResultList with regression metrics.
}
\examples{
Y = runif(100)
Y_hat = Y + 0.5*runif(100)
RegMetrics(Y = Y, Y_hat = Y_hat, MetricsNames = RegressionMetricsNames())
}
