% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{RegressionMetricsNames}
\alias{RegressionMetricsNames}
\title{Regression Metrics}
\usage{
RegressionMetricsNames()
}
\value{
character string vector with all possible regression metrics.
}
\description{
Returns a character string vector containning all regression metrics.
}
\examples{
RegressionMetricsNames()
}
