% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{ridge_regression}
\alias{ridge_regression}
\title{Ridge Regression}
\usage{
ridge_regression(X_train, Y_train, X_test, lambda)
}
\arguments{
\item{X_train}{A Matrix of trainning observations.}

\item{Y_train}{A numeric vector of classes or values of the trainning observations.}

\item{X_test}{A Matrix of testing observations.}

\item{lambda}{The penalty parameter.}
}
\value{
predicted values
}
\description{
Performs the ridge linear regression model.
}
\examples{
X <- as.matrix(cbind(runif(n = 100), runif(n = 100)))
Y <- 3*X[, 1] - 2.5*X[, 2] + 0.2 * runif(100)
pos <- sample(100, 70)
X_train <- X[pos, ]
X_test <- X[-pos, ]
Y_train <- Y[pos]
Y_test <- Y[-pos]
lambda <- 0.25
Y_predicted <- ridge_regression(X_train = X_train, Y_train = Y_train, X_test = X_test, lambda = lambda)
plot(x = (1:100)[-pos], y = Y_test, col = 'red')
points(x = (1:100)[-pos], y = Y_predicted, col = 'blue')
}
