% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tests.R
\name{RunSingleTest}
\alias{RunSingleTest}
\title{Run Single Test}
\usage{
RunSingleTest(testFunc, paramList, task, dataSetNames, metrics, nTestsPerParam,
  splitPerc, verbose = TRUE)
}
\description{
Runs single test
}
