% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tests.R
\name{RunTests}
\alias{RunTests}
\title{Run Tests}
\usage{
RunTests(cmpTestsFuncsList = cmpTestsFuncsList, task = task,
  dataSetNames = dataSetNames, metrics = NA, nTestsPerParam = 10,
  splitPerc = 0.7, verbose = TRUE)
}
\arguments{
\item{cmpTestsFuncsList}{is a list of}

\item{task}{A character string with 'MultClass' or 'BinClas'}

\item{dataSetNames}{A character string vector with valid dataset names, for options check}

\item{metrics}{character string vector with all testing metrics or NA for all available metrics}

\item{nTestsPerParam}{FOOO}

\item{splitPerc}{TODO}

\item{verbose}{TRUE/FALSE value for printing partial test}
}
\value{
TODO
}
\description{
Runs the tests given the methods and datasets
}
\examples{
cmpTestsFuncsList <- GetAllMultClassAlgo()
task <- 'MultClass'
dataSetNames <- c('Iris', 'PimaIndiansDiabetes')
myResult <- RunTests(cmpTestsFuncsList = GetAllMultClassAlgo(),
                     task = 'MultClass',
                     dataSetNames = c('Iris', 'PimaIndiansDiabetes'))

cmpTestsFuncsList <- GetAllRegressionAlgo()
task <- 'Regression'
dataSetNames <- GetDataSetsNames(task = task)
myResult <- RunTests(cmpTestsFuncsList = cmpTestsFuncsList,
                    task = task,
                    dataSetNames = dataSetNames)
}
