% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{squareConfusionTable}
\alias{squareConfusionTable}
\title{Square Confusion Matrix}
\usage{
squareConfusionTable(Y, Y_hat)
}
\arguments{
\item{Y}{A numeric vector for the ground truth labels}

\item{Y_hat}{A numeric vector for the predicted Labels}
}
\value{
A confusion table
}
\description{
Returns a square confusion matrix
}
\examples{
squareConfusionTable(Y = sample(1:2, size = 10, replace = TRUE), Y_hat = rep(1, 10))
}
