% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{GetMetrics}
\alias{GetMetrics}
\title{Get Metrics}
\usage{
GetMetrics(task)
}
\arguments{
\item{A}{valid task, can be 'MultClass' or 'BinClass'.}
}
\value{
All metrics names for that taks
}
\description{
Get all possible task names
}
\examples{
GetPossibleTasks()
}
