% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Avaliation.R
\name{hypotessesTesting}
\alias{hypotessesTesting}
\title{Hypoteses Testing}
\usage{
hypotessesTesting(resultTable, testingMetric, method = "friedman",
  reductionMethod = "mean", reductionOrder = "max", alpha = 0.05)
}
\arguments{
\item{resultTable}{Table resulted from a MLAT::RunTests() call}

\item{testingMetric}{A decision metric, must have been selected in MLAT::RunTests() function call.}

\item{reductionMethod}{A reduction method, can be 'mean' or 'median'.}

\item{reductionOrder}{The reduction Order of selected values, max or min.}

\item{alpha}{Test significance level.}
}
\value{
A list of type hypotesesTest.
}
\description{
Hypoteses Testing

Performs Statistical Tests
}
\examples{
cmpTestsFuncsList <- GetAllMultClassAlgo()
task <- 'MultClass'
dataSetNames <- MLAT::GetDataSetsNames(task = task)
myResult <- RunTests(cmpTestsFuncsList = GetAllMultClassAlgo(),
                     task = task,
                     dataSetNames = dataSetNames)
hypotessesTesting(resultTable = myResult,
                 testingMetric = 'F1 Macro',
                 method = 'friedman',
                 reductionMethod = 'mean',
                 reductionOrder = 'max',
                 alpha = 0.05)
}
